공지사항 total값 수정

+++ src/main/java/com/green/secondproject/notice/model/NoticeListVo.java
@@ -11,6 +11,6 @@ import java.util.List;
 public class NoticeListVo {
     private List<NoticeTotalVo> list;
     private Long total;
-    private Long totalPage;
+    private int totalPage;
 
 }
